grid: run celery beat in celeryworker itself - add cmd to run beat inside worker-start.sh - remove beat-start.sh and beat-start-reload.sh - define queue of periodic tasks to be main-queue
