+++ packages/grid/backend/grid/periodic_tasks.py
@@ -58,12 +58,3 @@ def cleanup_incomplete_uploads_from_blob_store() -> bool:
             )
 
     return True
-
-
-@celery_app.on_after_configure.connect
-def setup_periodic_task(sender, **kwargs) -> None:  # type: ignore
-    celery_app.add_periodic_task(
-        3600,  # Run every hour
-        cleanup_incomplete_uploads_from_blob_store.s(),
-        name="Clean incomplete uploads in Seaweed",
-    )
